Fixed a bug where all players were being deleted
There's some code that detects players changing their name or leaving
the FC, it's supposed to remove them from the database. There was a bug
in the error-handling when loading the FC list where if the lodestone
timed out, it would think there are no FC members and mark everyone for
deletion. Then on the next periodic scan, it would get the list again
and add everyone back into the list, but as visible=false by default